full refactor, completely decoupled components...
