+++ internal/collector/metric/config.go
@@ -5,7 +5,6 @@ import (
 	"github.com/turbolytics/latte/internal/metric"
 	"github.com/turbolytics/latte/internal/schedule"
 	"github.com/turbolytics/latte/internal/sink"
-	"github.com/turbolytics/latte/internal/sink/type"
 	"github.com/turbolytics/latte/internal/source"
 	"github.com/turbolytics/latte/internal/state"
 	"go.uber.org/zap"
@@ -48,6 +47,7 @@ func (c Config) GetSource() source.Config {
 	return c.Source
 }
 
+/*
 func (c Config) GetSinks() []_type.Sinker {
 	var ss []_type.Sinker
 	for _, s := range c.Sinks {
@@ -55,6 +55,7 @@ func (c Config) GetSinks() []_type.Sinker {
 	}
 	return ss
 }
+*/
 
 func (c Config) CollectorName() string {
 	return c.Name
@@ -74,6 +75,7 @@ func ConfigWithLogger(l *zap.Logger) ConfigOption {
 	}
 }
 
+/*
 // initSinks initializes all the outputs
 func initSinks(c *Config) error {
 	for k, v := range c.Sinks {
@@ -84,6 +86,7 @@ func initSinks(c *Config) error {
 	}
 	return nil
 }
+*/
 
 func defaults(c *Config) error {
 	(&c.Source).SetDefaults()
@@ -136,12 +139,15 @@ func NewConfig(raw []byte, opts ...ConfigOption) (*Config, error) {
 		return nil, err
 	}
 
-	if err := conf.Source.Init(conf.logger, conf.validate); err != nil {
-		return nil, err
-	}
+	/*
+		if err := conf.Source.Init(conf.logger, conf.validate); err != nil {
+			return nil, err
+		}
 
-	if err := initSinks(&conf); err != nil {
-		return nil, err
-	}
+		if err := initSinks(&conf); err != nil {
+			return nil, err
+		}
+
+	*/
 	return &conf, nil
 }
