+++ internal/sink/kafka/kafka.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"github.com/mitchellh/mapstructure"
 	"github.com/segmentio/kafka-go"
-	"github.com/turbolytics/latte/internal/sink/type"
+	"github.com/turbolytics/latte/internal/sink"
 )
 
 type config struct {
@@ -22,8 +22,8 @@ func (k *Kafka) Close() error {
 	return k.writer.Close()
 }
 
-func (k *Kafka) Type() _type.Type {
-	return _type.TypeKafka
+func (k *Kafka) Type() sink.Type {
+	return sink.TypeKafka
 }
 
 func (k *Kafka) Write(bs []byte) (int, error) {
