--- internal/sink/type/sinker.go
@@ -1,18 +1,0 @@
-package _type
-
-type Type string
-
-const (
-	TypeConsole Type = "console"
-	TypeHTTP    Type = "http"
-	TypeKafka   Type = "kafka"
-	TypeFile    Type = "file"
-)
-
-// Sinker is responsible for sinking
-// TODO - Starting with an io.Writer for right now.
-type Sinker interface {
-	Write([]byte) (int, error)
-	Close() error
-	Type() Type
-}
